+++ src/io.rs
@@ -134,7 +134,6 @@ pub trait Serialize: Sized {
 }
 
 use crate::{Error, Geometry, Result};
-use raw_assets::{extension, FileFormat};
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
@@ -149,7 +148,7 @@ impl Deserialize for crate::Texture2D {
             return img::deserialize_img(&path, &parse_data_url(path.to_str().unwrap())?);
         }
 
-        match raw_assets.guess_format(&path)? {
+        match FileFormat::guess(raw_assets, &path)? {
             FileFormat::Svg => {
                 #[cfg(not(feature = "svg"))]
                 return Err(Error::FeatureMissing("svg".to_string()));
@@ -194,8 +193,7 @@ impl Serialize for crate::Texture2D {
 impl Deserialize for crate::Scene {
     fn deserialize(path: impl AsRef<Path>, raw_assets: &mut RawAssets) -> Result<Self> {
         let path = raw_assets.match_path(path.as_ref())?;
-        let ext = raw_assets.guess_format(&path)?;
-        match ext {
+        match FileFormat::guess(raw_assets, &path)? {
             FileFormat::Gltf => {
                 #[cfg(not(feature = "gltf"))]
                 return Err(Error::FeatureMissing("gltf".to_string()));
@@ -268,7 +266,7 @@ impl Serialize for crate::Scene {
 impl Deserialize for crate::VoxelGrid {
     fn deserialize(path: impl AsRef<Path>, raw_assets: &mut RawAssets) -> Result<Self> {
         let path = raw_assets.match_path(path.as_ref())?;
-        match raw_assets.guess_extension(&path)? {
+        match FileFormat::guess(raw_assets, &path)? {
             FileExtension::Vol => {
                 #[cfg(not(feature = "vol"))]
                 return Err(Error::FeatureMissing("vol".to_string()));
@@ -338,7 +336,7 @@ fn dependencies(raw_assets: &RawAssets) -> Vec<PathBuf> {
     #[allow(unused_mut)]
     let mut dependencies = HashSet::new();
     for (path, _) in raw_assets.iter() {
-        match raw_assets.guess_format(path) {
+        match FileFormat::guess(raw_assets, &path) {
             Ok(FileFormat::Gltf) => {
                 #[cfg(feature = "gltf")]
                 dependencies.extend(gltf::dependencies(raw_assets, path));
@@ -383,4 +381,96 @@ fn parse_data_url(path: &str) -> Result<Vec<u8>> {
     }
     #[cfg(not(feature = "data-url"))]
     Err(Error::FeatureMissing("data-url".to_string()))
+}
+
+enum FileFormat {
+    Gltf,
+    Obj,
+    Mtl,
+    Stl,
+    Fbx,
+    Pcd,
+    ThreeMf,
+    Png,
+    Jpeg,
+    Gif,
+    Bmp,
+    Tga,
+    Tiff,
+    Hdr,
+    WebP,
+    Svg,
+    Vol,
+}
+
+impl FileFormat {
+    fn guess(raw_assets: &RawAssets, path: &Path) -> Result<Self> {
+        if let Ok(bytes) = raw_assets.get(path) {
+            if bytes.starts_with(b"glTF") {
+                return Ok(Self::Gltf);
+            }
+            if bytes.starts_with(b"Kaydara FBX Binary") {
+                return Ok(Self::Fbx);
+            }
+            if bytes.starts_with(&[0x50, 0x4B, 0x03, 0x04]) {
+                return Ok(Self::ThreeMf);
+            }
+            if bytes.starts_with(b"# .PCD") || bytes.starts_with(b"VERSION") {
+                return Ok(Self::Pcd);
+            }
+            if bytes.starts_with(b"solid ") {
+                return Ok(Self::Stl);
+            }
+            let trimmed = &bytes[bytes
+                .iter()
+                .position(|b| !b.is_ascii_whitespace())
+                .unwrap_or(0)..];
+            if trimmed.starts_with(b"<svg") || trimmed.starts_with(b"<?xml") {
+                return Ok(Self::Svg);
+            }
+            #[cfg(feature = "image")]
+            match image::guess_format(bytes) {
+                Ok(image::ImageFormat::Png) => return Ok(Self::Png),
+                Ok(image::ImageFormat::Jpeg) => return Ok(Self::Jpeg),
+                Ok(image::ImageFormat::Gif) => return Ok(Self::Gif),
+                Ok(image::ImageFormat::Bmp) => return Ok(Self::Bmp),
+                Ok(image::ImageFormat::Tga) => return Ok(Self::Tga),
+                Ok(image::ImageFormat::Tiff) => return Ok(Self::Tiff),
+                Ok(image::ImageFormat::Hdr) => return Ok(Self::Hdr),
+                Ok(image::ImageFormat::WebP) => return Ok(Self::WebP),
+                _ => {}
+            }
+        }
+
+        match extension(path).as_str() {
+            "gltf" | "glb" => return Ok(Self::Gltf),
+            "obj" => return Ok(Self::Obj),
+            "stl" => return Ok(Self::Stl),
+            "fbx" => return Ok(Self::Fbx),
+            "pcd" => return Ok(Self::Pcd),
+            "mtl" => return Ok(Self::Mtl),
+            "3mf" => return Ok(Self::ThreeMf),
+            "svg" => return Ok(Self::Svg),
+            "vol" => return Ok(Self::Vol),
+            "png" => return Ok(Self::Png),
+            "jpg" | "jpeg" => return Ok(Self::Jpeg),
+            "gif" => return Ok(Self::Gif),
+            "bmp" => return Ok(Self::Bmp),
+            "tga" => return Ok(Self::Tga),
+            "tiff" | "tif" => return Ok(Self::Tiff),
+            "hdr" => return Ok(Self::Hdr),
+            "webp" => return Ok(Self::WebP),
+            _ => {}
+        }
+
+        Err(Error::FailedToGuessFileExtension(
+            path.to_str().unwrap().to_string(),
+        ))
+    }
+}
+
+fn extension(path: &Path) -> String {
+    path.extension()
+        .map(|e| e.to_str().unwrap().to_ascii_lowercase())
+        .unwrap_or("".to_string())
 }
