+++ src/io/raw_assets.rs
@@ -5,98 +5,6 @@ use crate::{
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
-pub(crate) enum FileFormat {
-    Gltf,
-    Obj,
-    Mtl,
-    Stl,
-    Fbx,
-    Pcd,
-    ThreeMf,
-    Png,
-    Jpeg,
-    Gif,
-    Bmp,
-    Tga,
-    Tiff,
-    Hdr,
-    WebP,
-    Svg,
-    Vol,
-}
-
-impl FileFormat {
-    fn guess(raw_assets: &RawAssets, path: &Path) -> Result<Self> {
-        if let Ok(bytes) = raw_assets.get(path) {
-            if bytes.starts_with(b"glTF") {
-                return Ok(Self::Gltf);
-            }
-            if bytes.starts_with(b"Kaydara FBX Binary") {
-                return Ok(Self::Fbx);
-            }
-            if bytes.starts_with(&[0x50, 0x4B, 0x03, 0x04]) {
-                return Ok(Self::ThreeMf);
-            }
-            if bytes.starts_with(b"# .PCD") || bytes.starts_with(b"VERSION") {
-                return Ok(Self::Pcd);
-            }
-            if bytes.starts_with(b"solid ") {
-                return Ok(Self::Stl);
-            }
-            let trimmed = &bytes[bytes
-                .iter()
-                .position(|b| !b.is_ascii_whitespace())
-                .unwrap_or(0)..];
-            if trimmed.starts_with(b"<svg") || trimmed.starts_with(b"<?xml") {
-                return Ok(Self::Svg);
-            }
-            #[cfg(feature = "image")]
-            match image::guess_format(bytes) {
-                Ok(image::ImageFormat::Png) => return Ok(Self::Png),
-                Ok(image::ImageFormat::Jpeg) => return Ok(Self::Jpeg),
-                Ok(image::ImageFormat::Gif) => return Ok(Self::Gif),
-                Ok(image::ImageFormat::Bmp) => return Ok(Self::Bmp),
-                Ok(image::ImageFormat::Tga) => return Ok(Self::Tga),
-                Ok(image::ImageFormat::Tiff) => return Ok(Self::Tiff),
-                Ok(image::ImageFormat::Hdr) => return Ok(Self::Hdr),
-                Ok(image::ImageFormat::WebP) => return Ok(Self::WebP),
-                _ => {}
-            }
-        }
-
-        match extension(path).as_str() {
-            "gltf" | "glb" => return Ok(Self::Gltf),
-            "obj" => return Ok(Self::Obj),
-            "stl" => return Ok(Self::Stl),
-            "fbx" => return Ok(Self::Fbx),
-            "pcd" => return Ok(Self::Pcd),
-            "mtl" => return Ok(Self::Mtl),
-            "3mf" => return Ok(Self::ThreeMf),
-            "svg" => return Ok(Self::Svg),
-            "vol" => return Ok(Self::Vol),
-            "png" => return Ok(Self::Png),
-            "jpg" | "jpeg" => return Ok(Self::Jpeg),
-            "gif" => return Ok(Self::Gif),
-            "bmp" => return Ok(Self::Bmp),
-            "tga" => return Ok(Self::Tga),
-            "tiff" | "tif" => return Ok(Self::Tiff),
-            "hdr" => return Ok(Self::Hdr),
-            "webp" => return Ok(Self::WebP),
-            _ => {}
-        }
-
-        Err(Error::FailedToGuessFileExtension(
-            path.to_str().unwrap().to_string(),
-        ))
-    }
-}
-
-pub(crate) fn extension(path: &Path) -> String {
-    path.extension()
-        .map(|e| e.to_str().unwrap().to_ascii_lowercase())
-        .unwrap_or("".to_string())
-}
-
 ///
 /// Contains raw assets which are usually generated by either the [load](crate::io::load)/[load_async](crate::io::load_async) functions or the [Serialize::serialize](crate::io::Serialize::serialize) function.
 /// Can also be constructed manually using the [RawAssets::new] and [RawAssets::insert].
@@ -153,10 +61,6 @@ impl RawAssets {
     ///
     pub fn get(&self, path: impl AsRef<Path>) -> Result<&[u8]> {
         Ok(self.0.get(&self.match_path(path.as_ref())?).unwrap())
-    }
-
-    pub(crate) fn guess_format(&self, path: &Path) -> Result<FileFormat> {
-        FileFormat::guess(self, path)
     }
 
     pub(crate) fn match_path(&self, path: &Path) -> Result<PathBuf> {
